[alpaka] Move accelerator-independent headers outside of the alpaka directories
