+++ src/alpaka/plugin-PixelTriplets/CAConstants.h
