+++ src/alpaka/plugin-PixelTriplets/alpaka/GPUCACell.h
@@ -17,7 +17,7 @@
 #include "AlpakaDataFormats/PixelTrackAlpaka.h"
 #include "AlpakaDataFormats/TrackingRecHit2DAlpaka.h"
 
-#include "CAConstants.h"
+#include "../CAConstants.h"
 #include "../CircleEq.h"
 
 namespace ALPAKA_ACCELERATOR_NAMESPACE {
