+++ src/alpaka/plugin-PixelTriplets/alpaka/HelixFitOnGPU.h
@@ -5,7 +5,7 @@
 #include "AlpakaDataFormats/PixelTrackAlpaka.h"
 #include "AlpakaDataFormats/TrackingRecHit2DAlpaka.h"
 
-#include "CAConstants.h"
+#include "../CAConstants.h"
 #include "FitResult.h"
 
 namespace ALPAKA_ACCELERATOR_NAMESPACE {
